api/requirements.txt
fastapi
python-dotenv
azure-functions
geopy
pymongo
pydantic[email]